refactor(conversation): extract ConversationHitlService
R3 step 1. ConversationService: 2,735 -> 1,515 lines, and with it no
FileLength Checkstyle violation is left anywhere in the codebase — both
monoliths this plan set out to decompose are now under the limit.

The HITL cluster was the back ~43% of the class and had almost nothing to do
with the front. ConversationService is "run a turn"; this is "a turn stopped
and a person is deciding what happens next" — cancel/resume, per-tool
decision validation, the no-progress guard, timeout scheduling and bookmarks,
effective-policy resolution, and the compliance audit trail for all of it.

Resume re-enters the facade deliberately: applying a verdict means running
the rest of the turn, and that has to be the same say/step machinery a normal
turn uses. A resume-shaped second copy of turn execution is how post-approval
behaviour drifts from ordinary behaviour unnoticed. Hence the back-reference
for the eight facade members the cluster calls, same as GroupHitlCoordinator
in R1.

One real trap, caught by the tests: the collaborator was first built in a
@PostConstruct, which never fires for the ~34 test classes that construct
this service with `new` — 12 tests failed with NPEs. Moved into the
constructor, which is exactly what plan rule 3.0-4 exists to prevent.
Everything it needs is a constructor parameter or field initializer.

Bare-token sweep first: resumeConversation (118 refs), cancelConversation
(83), listPendingApprovals (46), CancelOutcome (27), plus reflection-reached
internals. All keep declared delegators, and the front half's own calls into
the cluster now route through the extracted service.

1,355 tests green.
